Overview: Rename overviewView to messageBoard
